+++ src/core/Entity.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include "Components.h"
+
+struct Entity
+{
+    int id;
+    TagComponent tag;
+    TransformComponent transform;
+    MaterialComponent material;
+};