+++ src/editor/InspectorPanel.h
@@ -3,19 +3,15 @@
 #include "EditorPanel.h"
 
 struct SelectionState;
+class Scene;
 
 class InspectorPanel : public EditorPanel
 {
 public:
-    InspectorPanel(SelectionState *selection);
+    InspectorPanel(SelectionState *selection, Scene *scene);
     void OnImGuiRender(float dt) override;
 
 private:
     SelectionState *m_selection;
-
-    float m_position[3];
-    float m_rotation[3];
-    float m_scale[3];
-    float m_color[4];
-    bool m_active;
+    Scene *m_scene;
 };