+++ src/editor/SceneHierarchyPanel.h
@@ -3,13 +3,15 @@
 #include "EditorPanel.h"
 
 struct SelectionState;
+class Scene;
 
 class SceneHierarchyPanel : public EditorPanel
 {
 public:
-    SceneHierarchyPanel(SelectionState *selection);
+    SceneHierarchyPanel(SelectionState *selection, Scene *scene);
     void OnImGuiRender(float dt) override;
 
 private:
     SelectionState *m_selection;
+    Scene *m_scene;
 };